Refactor: Move get_positions to its own file
